AsyncAPI rename, file upload and other changes
