--- modules/apigee_async_doc/apigee_async_doc.libraries.yml
@@ -1,23 +1,0 @@
-# Third-party library (CDN).
-reactjs:
-  remote: https://reactjs.org
-  version: 17.0.2
-  license:
-    name: MIT
-    url: https://github.com/facebook/react/blob/main/LICENSE
-    gpl-compatible: true
-  js:
-    https://unpkg.com/react@17.0.2/umd/react.development.js: {type: external, minified: true}
-    https://unpkg.com/react-dom@17.0.2/umd/react-dom.development.js: {type: external, minified: true}
-
-# Third-party library (CDN).
-asyncapi:
-  remote: https://www.asyncapi.com/
-  version: 0.24.19
-  license:
-    name: Apache License 2.0
-    url: https://github.com/asyncapi/asyncapi-react/blob/master/LICENSE
-    gpl-compatible: true
-  js:
-    https://unpkg.com/@webcomponents/webcomponentsjs@2.6.0/webcomponents-bundle.js: {type: external, minified: true}
-    https://unpkg.com/@asyncapi/web-component@0.24.19/lib/asyncapi-web-component.js: {type: external, minified: true}
