--- modules/apigee_async_doc/apigee_async_doc.links.action.yml
@@ -1,7 +1,0 @@
-node.async_doc.add_form:
-  route_name: node.add
-  title: 'AsyncAPI Doc'
-  route_parameters:
-    node_type: 'async_doc'
-  appears_on:
-    - view.api_catalog_admin.page_1
