--- modules/apigee_async_doc/apigee_async_doc.module
@@ -1,42 +1,0 @@
-<?php
-
-/**
- * @file
- * Copyright 2022 Google Inc.
- *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License version 2 as published by the
- * Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
- * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public
- * License for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc., 51
- * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- */
-
-/**
- * Primary module hooks for Apigee Async Doc module.
- *
- * @DCG
- * This file is no longer required in Drupal 8.
- * @see https://www.drupal.org/node/2217931
- */
-
- /**
-  * Implements hook_theme().
-  */
- function apigee_async_doc_theme($existing, $type, $theme, $path) {
-   return [
-     'apigee_async_doc_file_link_field_item' => [
-       'variables' => [
-         'field_name' => NULL,
-         'delta' => NULL,
-         'link' => NULL,
-       ],
-     ],
-   ];
- }
