--- modules/apigee_async_doc/apigee_async_doc.services.yml
@@ -1,17 +1,0 @@
-services:
-  apigee_async_doc.event_subscriber:
-    class: Drupal\apigee_async_doc\EventSubscriber\ApigeeAsyncDocSubscriber
-    arguments: ['@messenger']
-    tags:
-      - { name: event_subscriber }
-
-  apigee_async_doc.uninstall_validator:
-    class: Drupal\apigee_async_doc\ApigeeAsyncDocUninstallValidator
-    tags:
-      - { name: module_install.uninstall_validator }
-    arguments: ['@entity_type.manager', '@string_translation']
-
-  apigee_async_doc.breadcrumb:
-    class: Drupal\apigee_async_doc\ApigeeAsyncDocBreadcrumbBuilder
-    tags:
-      - { name: breadcrumb_builder, priority: 1000 }
