--- modules/apigee_async_doc/config/install/field.field.node.async_doc.body.yml
@@ -1,22 +1,0 @@
-langcode: en
-status: true
-dependencies:
-  config:
-    - field.storage.node.body
-    - node.type.async_doc
-  module:
-    - text
-id: node.async_doc.body
-field_name: body
-entity_type: node
-bundle: async_doc
-label: Description
-description: ''
-required: false
-translatable: true
-default_value: {  }
-default_value_callback: ''
-settings:
-  display_summary: true
-  required_summary: false
-field_type: text_with_summary
