--- modules/apigee_async_doc/config/install/field.storage.node.field_async_file_link.yml
@@ -1,21 +1,0 @@
-langcode: en
-status: true
-dependencies:
-  enforced:
-    module:
-      - apigee_async_doc
-  module:
-    - file_link
-    - node
-id: node.field_async_file_link
-field_name: field_async_file_link
-entity_type: node
-type: file_link
-settings: {  }
-module: file_link
-locked: false
-cardinality: 1
-translatable: true
-indexes: {  }
-persist_with_no_fields: false
-custom_storage: false
