--- modules/apigee_async_doc/config/install/node.type.async_doc.yml
@@ -1,16 +1,0 @@
-langcode: en
-dependencies:
-  module:
-    - apigee_api_catalog
-    - apigee_async_doc
-  enforced:
-    module:
-      - apigee_async_doc
-type: async_doc
-name: 'AsyncAPI Doc'
-description: 'Use <em>AsyncAPI Doc</em> to document AsyncAPI.'
-help: ''
-new_revision: false
-display_submitted: false
-preview_mode: 0
-status: true
