--- modules/apigee_async_doc/config/optional/core.entity_view_display.node.async_doc.default.yml
@@ -1,44 +1,0 @@
-langcode: en
-status: true
-dependencies:
-  config:
-    - field.field.node.async_doc.body
-    - field.field.node.async_doc.field_api_product
-    - field.field.node.async_doc.field_async_file_link
-    - node.type.async_doc
-  module:
-    - file_link
-    - text
-    - user
-id: node.async_doc.default
-targetEntityType: node
-bundle: async_doc
-mode: default
-content:
-  body:
-    label: hidden
-    type: text_default
-    weight: 1
-    settings: {  }
-    third_party_settings: {  }
-    region: content
-  field_async_file_link:
-    weight: 3
-    label: hidden
-    settings:
-      trim_length: 80
-      url_only: true
-      url_plain: false
-      rel: ''
-      target: ''
-      format_size: true
-    third_party_settings: {  }
-    type: apigee_async_doc_async
-    region: content
-  links:
-    weight: 0
-    region: content
-    settings: {  }
-    third_party_settings: {  }
-hidden:
-  field_api_product: true
