+++ modules/apigee_asyncapi_doc/apigee_asyncapi_doc.libraries.yml
@@ -1,0 +1,24 @@
+# Third-party library (CDN).
+reactjs:
+  remote: https://reactjs.org
+  version: 17.0.2
+  license:
+    name: MIT
+    url: https://github.com/facebook/react/blob/main/LICENSE
+    gpl-compatible: true
+  js:
+    https://unpkg.com/react@17.0.2/umd/react.production.min.js: {type: external, minified: true}
+    https://unpkg.com/react-dom@17.0.2/umd/react-dom.production.min.js: {type: external, minified: true}
+
+# Third-party library (CDN).
+# When updating the version, apigee-async-doc.css should be updated as well.
+asyncapi:
+  remote: https://www.asyncapi.com/
+  version: 0.24.19
+  license:
+    name: Apache License 2.0
+    url: https://github.com/asyncapi/asyncapi-react/blob/master/LICENSE
+    gpl-compatible: true
+  js:
+    https://unpkg.com/@webcomponents/webcomponentsjs@2.6.0/webcomponents-bundle.js: {type: external, minified: true}
+    https://unpkg.com/@asyncapi/web-component@0.24.19/lib/asyncapi-web-component.js: {type: external, minified: true}
