+++ modules/apigee_asyncapi_doc/apigee_asyncapi_doc.links.action.yml
@@ -1,0 +1,7 @@
+node.asyncapi_doc.add_form:
+  route_name: node.add
+  title: 'AsyncAPI'
+  route_parameters:
+    node_type: 'asyncapi_doc'
+  appears_on:
+    - view.api_catalog_admin.page_1
