+++ modules/apigee_asyncapi_doc/apigee_asyncapi_doc.routing.yml
@@ -1,0 +1,7 @@
+apigee_asyncapi_doc.confirm_remove_views_references:
+  path: '/admin/modules/uninstall/apigee-async-doc/confirm-remove-views-references'
+  defaults:
+    _title: 'Confirm remove views references'
+    _form: 'Drupal\apigee_asyncapi_doc\Form\ConfirmRemoveViewsReferencesForm'
+  requirements:
+    _permission: 'administer modules'
