+++ modules/apigee_asyncapi_doc/config/install/field.field.node.asyncapi_doc.body.yml
@@ -1,0 +1,22 @@
+langcode: en
+status: true
+dependencies:
+  config:
+    - field.storage.node.body
+    - node.type.asyncapi_doc
+  module:
+    - text
+id: node.asyncapi_doc.body
+field_name: body
+entity_type: node
+bundle: asyncapi_doc
+label: Description
+description: ''
+required: false
+translatable: true
+default_value: {  }
+default_value_callback: ''
+settings:
+  display_summary: true
+  required_summary: false
+field_type: text_with_summary
