+++ modules/apigee_asyncapi_doc/config/install/field.storage.node.field_asyncapi_spec.yml
@@ -1,0 +1,25 @@
+langcode: en
+status: true
+dependencies:
+  enforced:
+    module:
+      - apigee_asyncapi_doc
+  module:
+    - file
+    - node
+id: node.field_asyncapi_spec
+field_name: field_asyncapi_spec
+entity_type: node
+type: file
+settings:
+  target_type: file
+  display_field: false
+  display_default: false
+  uri_scheme: public
+module: file
+locked: false
+cardinality: 1
+translatable: false
+indexes: {  }
+persist_with_no_fields: false
+custom_storage: false
