+++ modules/apigee_asyncapi_doc/config/install/field.storage.node.field_asyncapi_spec_file_link.yml
@@ -1,0 +1,21 @@
+langcode: en
+status: true
+dependencies:
+  enforced:
+    module:
+      - apigee_asyncapi_doc
+  module:
+    - file_link
+    - node
+id: node.field_asyncapi_spec_file_link
+field_name: field_asyncapi_spec_file_link
+entity_type: node
+type: file_link
+settings: {  }
+module: file_link
+locked: false
+cardinality: 1
+translatable: false
+indexes: {  }
+persist_with_no_fields: false
+custom_storage: false
