+++ modules/apigee_asyncapi_doc/css/apigee-asyncapi-doc.css
@@ -1,0 +1,32 @@
+/**
+ * Copyright 2022 Google Inc.
+ *
+ * This program is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License version 2 as published by the
+ * Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public
+ * License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc., 51
+ * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ */
+
+ @import url(https://unpkg.com/@asyncapi/react-component@0.24.19/lib/styles/fiori.css);
+
+/* Customization as per Apigee */
+.asyncapi,
+.asyncapi__tag,
+.asyncapi__server-expand-icon:before,
+.asyncapi__error {
+  font-family: Roboto,sans-serif;
+}
+.asyncapi__code-body {
+  font-family: courier;
+}
+h1, h2, h3 {
+  font-weight: 300;
+}
