+++ modules/apigee_asyncapi_doc/templates/apigee-asyncapi-doc-file-link-field-item.html.twig
@@ -1,0 +1,38 @@
+{#
+/**
+  * @file apigee-async-doc-file-link-field-item.html.twig
+  * Copyright 2022 Google Inc.
+  *
+  * This program is free software; you can redistribute it and/or modify it under
+  * the terms of the GNU General Public License version 2 as published by the
+  * Free Software Foundation.
+  *
+  * This program is distributed in the hope that it will be useful, but WITHOUT
+  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+  * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public
+  * License for more details.
+  *
+  * You should have received a copy of the GNU General Public License along
+  * with this program; if not, write to the Free Software Foundation, Inc., 51
+  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+  */
+
+/**
+  * Default theme implementation for Apigee AsyncAPI Doc field items.
+  *
+  * Available variables:
+  * - field_name: The name of the field.
+  * - delta: The delta of the field item.
+  * - link: A link to the file.
+  *
+  * @ingroup themeable
+  */
+#}
+<div id="apigee-asyncapi-doc-{{ field_name }}-{{ delta }}">
+  {% set module_path = _self|split('/templates') %}
+  <i class="fas fa-expand fa-fw fa-3x">Expand</i>
+  <asyncapi-component
+    schemaUrl="{{ link }}"
+    cssImportPath="{{ url('<front>') }}{{ module_path[0] }}/css/apigee-asyncapi-doc.css">
+  </asyncapi-component>
+</div>
